+++ pybayes/stresses/__main__.py
@@ -1,8 +1,8 @@
-# Copyright (c) 2011 Matej Laitl <matej@laitl.cz>
+# Copyright (c) 2013 Matej Laitl <matej@laitl.cz>
 # Distributed under the terms of the GNU General Public License v2 or any
 # later version of the license, at your option.
 
-"""PyBayes test-suite runner. Used when user calls `python -m pybayes.tests"""
+"""PyBayes stress-suite runner. Used when user calls `python -m pybayes.stresses"""
 
 import unittest as ut