+++ ChangeLog.rst
@@ -12,7 +12,7 @@ Changes between 0.3 and 0.4
 * ParticleFilter.bayes() now ignores cond completely. Yell if you need it.
 * ParticleFilter lost last emp_pdf argument. Pass the same object as the init_pdf argument
   to achieve the same thing.
-* Test-suite no longer needs PyBayes to be installed, yay! (no privilege problems etc.)
+* Test- and stress-suite no longer need PyBayes to be installed. (no privilege problems etc.)
 * Build-system was rewritten so that it is no longer an ugly hack. .pxd and .py files are now
   installed along .so (.dll) files for interoperability and additional openness. Better parsing of
   setup.py arguments and custom parameters visible in the --help command.